+++ CHANGES.rst
@@ -10,6 +10,7 @@ Changelog
   of the pipeline number.
 - Fix TypeError: sequence item 0: expected str instance, int found
   while printing "Found multiple jobs"
+- ``--print-url`` for printing the job's URL instead of dumping the log.
 
 
 0.5.0 (2020-09-18)